Refactor[speedometer]: change module name
