+++ speedometer/speedometer_test.go
@@ -1,4 +1,4 @@
-package util
+package speedometer
 
 import (
 	"bytes"
